pyproject.toml
[project]
name = "tubes3-lo-siento"
version = "0.1.0"
description = "ATS CV Analyzer - Aplikasi parsing CV dengan algoritma pattern matching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pymupdf==1.24.2",
    "mysql-connector-python==9.3.0",
    "pyqt6==6.9.0",
    "python-dotenv==1.0.1",
]
